issue #472: check if s3 client configured before using it

+++ classes/local/store/s3/client.php
@@ -79,29 +79,72 @@ class client extends object_client_base {
         // it will be serialised, so re-retrive them now.
         $config = manager::get_objectfs_config();
         $this->set_client($config);
-        $this->client->registerStreamWrapper();
+        if ($this->is_functional()) {
+            $this->client->registerStreamWrapper();
+        }
     }
 
-    public function set_client($config) {
-        $options = array(
-            'region' => $config->s3_region,
-            'version' => AWS_API_VERSION
-        );
+    /**
+     * Check if the client is functional.
+     * @return bool
+     */
+    private function is_functional() {
+        return !is_null($this->client);
+    }
+
+    /**
+     * Check if the client configured properly.
+     *
+     * @param \stdClass $config Client config.
+     * @return bool
+     */
+    private function is_configured($config) {
+        if (empty($config->s3_bucket)) {
+            return false;
+        }
+
+        if (empty($config->s3_region)) {
+            return false;
+        }
 
         if (empty($config->s3_usesdkcreds)) {
-            $options['credentials'] = array('key' => $config->s3_key, 'secret' => $config->s3_secret);
+            if (empty($config->s3_key) && empty($config->s3_secret)) {
+                return false;
+            }
         }
 
-        if ($config->useproxy) {
-            $options['http'] = array('proxy' => $this->get_proxy_string());
-        }
+        return true;
+    }
 
-        // Support base_url config for aws api compatible endpoints.
-        if ($config->s3_base_url) {
-            $options['endpoint'] = $config->s3_base_url;
-        }
+    /**
+     * Set the client.
+     *
+     * @param \stdClass $config Client config.
+     */
+    public function set_client($config) {
+        if (!$this->is_configured($config)) {
+            $this->client = null;
+        } else {
+            $options = array(
+                'region' => $config->s3_region,
+                'version' => AWS_API_VERSION
+            );
 
-        $this->client = \Aws\S3\S3Client::factory($options);
+            if (empty($config->s3_usesdkcreds)) {
+                $options['credentials'] = array('key' => $config->s3_key, 'secret' => $config->s3_secret);
+            }
+
+            if ($config->useproxy) {
+                $options['http'] = array('proxy' => $this->get_proxy_string());
+            }
+
+            // Support base_url config for aws api compatible endpoints.
+            if ($config->s3_base_url) {
+                $options['endpoint'] = $config->s3_base_url;
+            }
+
+            $this->client = \Aws\S3\S3Client::factory($options);
+        }
     }
 
     /**
@@ -109,7 +152,7 @@ class client extends object_client_base {
      *
      */
     public function register_stream_wrapper() {
-        if ($this->get_availability()) {
+        if ($this->get_availability() && $this->is_functional()) {
             $this->client->registerStreamWrapper();
         } else {
             parent::register_stream_wrapper();
@@ -117,6 +160,10 @@ class client extends object_client_base {
     }
 
     private function get_md5_from_hash($contenthash) {
+        if (!$this->is_functional()) {
+            return false;
+        }
+
         try {
             $key = $this->get_filepath_from_hash($contenthash);
             $result = $this->client->headObject(array(
@@ -206,17 +253,23 @@ class client extends object_client_base {
         $connection->success = true;
         $connection->details = '';
 
-        try {
-            $this->client->headBucket(array('Bucket' => $this->bucket));
-        } catch (\Aws\S3\Exception\S3Exception $e) {
+        if (!$this->is_functional()) {
             $connection->success = false;
-            $connection->details = $this->get_exception_details($e);
-        } catch (\GuzzleHttp\Exception\InvalidArgumentException $e) {
-            $connection->success = false;
-            $connection->details = $this->get_exception_details($e);
-        } catch (\Aws\Exception\CredentialsException $e) {
-            $connection->success = false;
-            $connection->details = $this->get_exception_details($e);
+            $connection->details = '';
+        } else {
+            try {
+                $this->client->headBucket(array('Bucket' => $this->bucket));
+            } catch (\Aws\S3\Exception\S3Exception $e) {
+                $connection->success = false;
+                $connection->details = $this->get_exception_details($e);
+            } catch (\GuzzleHttp\Exception\InvalidArgumentException $e) {
+                $connection->success = false;
+                $connection->details = $this->get_exception_details($e);
+            } catch (\Aws\Exception\CredentialsException $e) {
+                $connection->success = false;
+                $connection->details = $this->get_exception_details($e);
+            }
+
         }
 
         return $connection;
@@ -235,49 +288,54 @@ class client extends object_client_base {
         $permissions->success = true;
         $permissions->messages = array();
 
-        try {
-            $result = $this->client->putObject(array(
-                            'Bucket' => $this->bucket,
-                            'Key' => $this->bucketkeyprefix . 'permissions_check_file',
-                            'Body' => 'test content'));
-        } catch (\Aws\S3\Exception\S3Exception $e) {
-            $details = $this->get_exception_details($e);
-            $permissions->messages[get_string('settings:writefailure', 'tool_objectfs') . $details] = 'notifyproblem';
+        if (!$this->is_functional()) {
             $permissions->success = false;
-        }
-
-        try {
-            $result = $this->client->getObject(array(
-                            'Bucket' => $this->bucket,
-                            'Key' => $this->bucketkeyprefix . 'permissions_check_file'));
-        } catch (\Aws\S3\Exception\S3Exception $e) {
-            $errorcode = $e->getAwsErrorCode();
-            // Write could have failed.
-            if ($errorcode !== 'NoSuchKey') {
+            $permissions->messages = array();
+        } else {
+            try {
+                $result = $this->client->putObject(array(
+                    'Bucket' => $this->bucket,
+                    'Key' => $this->bucketkeyprefix . 'permissions_check_file',
+                    'Body' => 'test content'));
+            } catch (\Aws\S3\Exception\S3Exception $e) {
                 $details = $this->get_exception_details($e);
-                $permissions->messages[get_string('settings:readfailure', 'tool_objectfs') . $details] = 'notifyproblem';
+                $permissions->messages[get_string('settings:writefailure', 'tool_objectfs') . $details] = 'notifyproblem';
                 $permissions->success = false;
             }
-        }
 
-        if ($testdelete) {
             try {
-                $result = $this->client->deleteObject(array('Bucket' => $this->bucket, 'Key' => $this->bucketkeyprefix . 'permissions_check_file'));
-                $permissions->messages[get_string('settings:deletesuccess', 'tool_objectfs')] = 'warning';
-                $permissions->success = false;
+                $result = $this->client->getObject(array(
+                    'Bucket' => $this->bucket,
+                    'Key' => $this->bucketkeyprefix . 'permissions_check_file'));
             } catch (\Aws\S3\Exception\S3Exception $e) {
                 $errorcode = $e->getAwsErrorCode();
-                // Something else went wrong.
-                if ($errorcode !== 'AccessDenied') {
+                // Write could have failed.
+                if ($errorcode !== 'NoSuchKey') {
                     $details = $this->get_exception_details($e);
-                    $permissions->messages[get_string('settings:deleteerror', 'tool_objectfs') . $details] = 'notifyproblem';
+                    $permissions->messages[get_string('settings:readfailure', 'tool_objectfs') . $details] = 'notifyproblem';
                     $permissions->success = false;
                 }
             }
-        }
 
-        if ($permissions->success) {
-            $permissions->messages[get_string('settings:permissioncheckpassed', 'tool_objectfs')] = 'notifysuccess';
+            if ($testdelete) {
+                try {
+                    $result = $this->client->deleteObject(array('Bucket' => $this->bucket, 'Key' => $this->bucketkeyprefix . 'permissions_check_file'));
+                    $permissions->messages[get_string('settings:deletesuccess', 'tool_objectfs')] = 'warning';
+                    $permissions->success = false;
+                } catch (\Aws\S3\Exception\S3Exception $e) {
+                    $errorcode = $e->getAwsErrorCode();
+                    // Something else went wrong.
+                    if ($errorcode !== 'AccessDenied') {
+                        $details = $this->get_exception_details($e);
+                        $permissions->messages[get_string('settings:deleteerror', 'tool_objectfs') . $details] = 'notifyproblem';
+                        $permissions->success = false;
+                    }
+                }
+            }
+
+            if ($permissions->success) {
+                $permissions->messages[get_string('settings:permissioncheckpassed', 'tool_objectfs')] = 'notifysuccess';
+            }
         }
 
         return $permissions;
